add formsmodule, update file names, add spark thee
