+++ angular/projects/spark-angular/src/lib/components/inputs/sprk-helpertext.stories.ts
@@ -3,17 +3,21 @@ import { SprkInputContainerModule } from './sprk-input-container/sprk-input-cont
 import { SprkInputModule } from '../../directives/inputs/sprk-input/sprk-input.module';
 import { SprkIconModule } from '../sprk-icon/sprk-icon.module';
 import { SprkLabelModule } from '../../directives/inputs/sprk-label/sprk-label.module';
-import { SprkFieldErrorModule } from '../../directives/inputs/sprk-field-error/sprk-field-error.module';
-import { SprkFieldErrorDirective } from '../../directives/inputs/sprk-field-error/sprk-field-error.directive';
-import { SprkInputContainerComponent } from './sprk-input-container/sprk-input-container.component';
-import { SprkInputDirective } from '../../directives/inputs/sprk-input/sprk-input.directive';
 import { SprkLabelDirective } from '../../directives/inputs/sprk-label/sprk-label.directive';
+import { SprkInputDirective } from '../../directives/inputs/sprk-input/sprk-input.directive';
+import { SprkFieldErrorModule } from '../../directives/inputs/sprk-field-error/sprk-field-error.module';
+import { SprkHelperTextModule } from '../../directives/inputs/sprk-helper-text/sprk-helper-text.module';
+import { SprkInputContainerComponent } from './sprk-input-container/sprk-input-container.component';
+import { SprkHelperTextDirective } from '../../directives/inputs/sprk-helper-text/sprk-helper-text.directive';
+import { SprkFieldErrorDirective } from '../../directives/inputs/sprk-field-error/sprk-field-error.directive';
 import { markdownDocumentationLinkBuilder } from '../../../../../../../storybook-utilities/markdownDocumentationLinkBuilder';
+import { FormsModule } from '@angular/forms';
 
 export default {
-  title: 'Components/Input/Date',
+  title: 'Components/Input/Helper Text',
   component: SprkInputContainerComponent,
   subcomponents: {
+    SprkHelperTextDirective,
     SprkLabelDirective,
     SprkInputDirective,
     SprkFieldErrorDirective,
@@ -36,13 +40,8 @@ export default {
   parameters: {
     info: `
 ${markdownDocumentationLinkBuilder('input')}
-- The value of this field contains special characters (/)
-which you may need to remove before submitting the form.
-- Requires Additional Engineering:
-  - Validation
-  - Reformatting to date pattern MM/DD/YYYY - After valid
-  input, it should reformat to add commas and decimals
-  to the hundredth place.
+- Helper text must be placed below the Input and above the error container.
+- Helper text is marked with the \`sprkHelperText\` directive.
 `,
     docs: { iframeHeight: 200 },
   },
@@ -55,54 +54,69 @@ const modules = {
     SprkInputModule,
     SprkIconModule,
     SprkFieldErrorModule,
+    SprkHelperTextModule,
+    FormsModule,
   ],
 };
 
-export const dateInput = () => ({
+export const helperText = () => ({
   moduleMetadata: modules,
   template: `
     <sprk-input-container>
-      <label for="date" sprkLabel>Date Input (No Picker)</label>
+      <label for="helper-input" sprkLabel>
+        Text Input Label
+      </label>
       <input
-        id="date"
-        name="date_input"
-        type="text"
-        placeholder="MM/DD/YYYY"
-        [(ngModel)]="date_input"
-        #dateInput="ngModel"
+        id="helper-input"
+        aria-describedby="helper-text"
         sprkInput
-      />
+        name="text_input"
+        type="text"
+        [(ngModel)]="text_input"
+        #textInput="ngModel"
+        idString="text-input-1"
+      >
+      <p sprkHelperText id="helper-text">
+        Optional helper text, used to clarify the field's intent.
+      </p>
     </sprk-input-container>
   `,
 });
 
-dateInput.storyName = 'Default';
+helperText.storyName = 'Default';
 
-dateInput.parameters = {
+helperText.parameters = {
   jest: [
     'sprk-input-container.component',
     'sprk-input.directive',
     'sprk-label.directive',
+    'sprk-helper-text.directive',
   ],
 };
 
-export const invalidDateInput = () => ({
+export const invalidHelperText = () => ({
   moduleMetadata: modules,
   template: `
     <sprk-input-container>
-      <label for="invalid-date" sprkLabel>Date Input (No Picker)</label>
+      <label for="invalid-helper-input" sprkLabel>
+        Text Input Label
+      </label>
       <input
-        id="invalid-date"
-        aria-describedby="date-error"
-        class="sprk-b-TextInput--error"
-        name="date_input"
-        type="text"
-        placeholder="MM/DD/YYYY"
-        [(ngModel)]="date_input"
-        #dateInput="ngModel"
+        id="invalid-helper-input"
+        aria-describedby="helper-text helper-error"
         sprkInput
-      />
-      <span sprkFieldError id="date-error">
+        name="text_input"
+        type="text"
+        [(ngModel)]="text_input"
+        #textInput="ngModel"
+        idString="text-input-1"
+        class="sprk-b-TextInput--error"
+        aria-invalid="true"
+      >
+      <p sprkHelperText id="helper-text">
+        Optional helper text, used to clarify the field's intent.
+      </p>
+      <span sprkFieldError id="helper-error">
         <sprk-icon
           iconName="exclamation-filled"
           additionalClasses="sprk-b-ErrorIcon"
@@ -114,42 +128,14 @@ export const invalidDateInput = () => ({
   `,
 });
 
-invalidDateInput.storyName = 'Invalid';
+invalidHelperText.storyName = 'With Error Text';
 
-invalidDateInput.parameters = {
+invalidHelperText.parameters = {
   jest: [
     'sprk-input-container.component',
     'sprk-input.directive',
     'sprk-label.directive',
+    'sprk-helper-text.directive',
     'sprk-field-error.directive',
-  ],
-};
-
-export const disabledDateInput = () => ({
-  moduleMetadata: modules,
-  template: `
-    <sprk-input-container>
-      <label for="disabled-date" sprkLabel isDisabled="true">Date Input (No Picker)</label>
-      <input
-        id="disabled-date"
-        disabled
-        name="date_input"
-        type="text"
-        placeholder="MM/DD/YYYY"
-        [(ngModel)]="date_input"
-        #dateInput="ngModel"
-        sprkInput
-      />
-    </sprk-input-container>
-  `,
-});
-
-disabledDateInput.storyName = 'Disabled';
-
-disabledDateInput.parameters = {
-  jest: [
-    'sprk-input-container.component',
-    'sprk-input.directive',
-    'sprk-label.directive',
   ],
 };
