+++ angular/projects/spark-angular/src/lib/components/inputs/sprk-phone.stories.ts
@@ -9,6 +9,7 @@ import { SprkInputDirective } from '../../directives/inputs/sprk-input/sprk-inpu
 import { SprkLabelDirective } from '../../directives/inputs/sprk-label/sprk-label.directive';
 import { SprkFieldErrorDirective } from '../../directives/inputs/sprk-field-error/sprk-field-error.directive';
 import { markdownDocumentationLinkBuilder } from '../../../../../../../storybook-utilities/markdownDocumentationLinkBuilder';
+import { FormsModule } from '@angular/forms';
 
 export default {
   title: 'Components/Input/Phone',
@@ -57,6 +58,7 @@ const modules = {
     SprkInputModule,
     SprkIconModule,
     SprkFieldErrorModule,
+    FormsModule,
   ],
 };
