+++ angular/projects/spark-angular/src/lib/components/inputs/sprk-search.stories.ts
@@ -11,6 +11,7 @@ import { SprkFieldErrorModule } from '../../directives/inputs/sprk-field-error/s
 import { SprkInputDirective } from '../../directives/inputs/sprk-input/sprk-input.directive';
 import { SprkFieldErrorDirective } from '../../directives/inputs/sprk-field-error/sprk-field-error.directive';
 import { markdownDocumentationLinkBuilder } from '../../../../../../../storybook-utilities/markdownDocumentationLinkBuilder';
+import { FormsModule } from '@angular/forms';
 
 export default {
   title: 'Components/Input/Search',
@@ -52,6 +53,7 @@ const modules = {
     SprkInputModule,
     SprkIconModule,
     SprkFieldErrorModule,
+    FormsModule,
   ],
 };
